Fix text alignment on storage pool type combo boxes
The storage pool type list does not have a left padding like the
other combo boxes in Kimchi, bringing the text very close to the UI
component's border.
The SCSI device list on the same window has the same alignment problem.

Apply a common CSS class to the combo boxes so they look like the other
objects in the application.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>